requirements-discord.txt
discord.py>=2.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
